fix: get things running
