+++ containers/ecr-viewer/src/app/services/labsService.tsx
@@ -1,3 +1,4 @@
+import "server-only";
 import React from "react";
 import { Bundle, Device, Observation, Organization, Reference } from "fhir/r4";
 import {
@@ -42,25 +43,6 @@ export interface LabReportElementData {
   diagnosticReportDataItems: AccordionItem[];
   organizationDisplayDataProps: DisplayDataProps[];
 }
-
-/**
- * Checks if a given list is of type LabReportElementData[].
- * Used to determine how to render lab results.
- * @param labResults - Object to be checked.
- * @returns True if the list is of type LabReportElementData[], false otherwise.
- */
-export const isLabReportElementDataList = (
-  labResults: DisplayDataProps[] | LabReportElementData[],
-): labResults is LabReportElementData[] => {
-  const asLabReportElementList = labResults as LabReportElementData[];
-  return (
-    asLabReportElementList &&
-    asLabReportElementList.length > 0 &&
-    asLabReportElementList[0].diagnosticReportDataItems !== undefined &&
-    asLabReportElementList[0].organizationId !== undefined &&
-    asLabReportElementList[0].organizationDisplayDataProps !== undefined
-  );
-};
 
 /**
  * Extracts an array of `Observation` resources from a given FHIR bundle based on a list of observation references.
